simplify "baseline" hack
The baseline/shift correction is not needed in modern edge/ff anymore (in fact, it had worsened the rendering in Chrome, as the "hack" had shifted everything down though the `baseline-shift`).

In fact, it appears as if only Safari does not properly handle the `dominant-baseline` property in SVG, so only it needs the correction.

see #3020, #7044